refactor: 🧹 Jules05: Product coherence improvements
Audit and remediation of system coherence:
- Harmonized Risk Management: Consolidated RiskEngine into RiskManager with unified 8-layer safety cascade.
- Module Boundary Alignment: Relocated FeatureEngineer to src/data/feature_engineering.py.
- Logic Consolidation: Updated main.py to use risk-manager-approved lot sizes.
- Cleanup: Removed redundant risk_engine.py and clarified package exports.
- Documentation: Formalized PRODUCT_COHERENCE_AUDIT.md.
