+++ src/core/__init__.py
@@ -1,7 +1,5 @@
 """Core configuration and settings."""
 
-from typing import TYPE_CHECKING
-
 from src.core.audit_log import AuditLogger, get_audit_logger
 from src.core.config import TradingConfig, get_config
 from src.core.decision_support import DecisionPacket, DecisionSupportSystem
@@ -9,23 +7,10 @@ from src.core.explainability import SignalExplainer, SignalExplanation
 from src.core.monitor import Monitor
 from src.core.profiler import profile
 
-if TYPE_CHECKING:
-    from src.core.feature_engineering import FeatureEngineer
-else:
-    # Lazy load FeatureEngineer to avoid early talib dependency
-    def __getattr__(name):
-        if name == "FeatureEngineer":
-            from src.core.feature_engineering import FeatureEngineer
-
-            return FeatureEngineer
-        raise AttributeError(f"module {__name__} has no attribute {name}")
-
-
 __all__ = [
     "AuditLogger",
     "DecisionPacket",
     "DecisionSupportSystem",
-    "FeatureEngineer",
     "Monitor",
     "SignalExplainer",
     "SignalExplanation",
