+++ src/trading/audited_risk_manager.py
@@ -9,11 +9,13 @@ License: MIT
 from __future__ import annotations
 
 import logging
-from typing import Optional
+from typing import Any, Dict, List, Optional
+
+import pandas as pd
 
 from src.core.audit_log import get_audit_logger
 from src.core.schemas import TradeSignal
-from src.trading.risk_manager import RiskManager
+from src.trading.risk_manager import RiskDecision, RiskManager
 
 logger = logging.getLogger(__name__)
 
@@ -24,25 +26,33 @@ class AuditedRiskManager(RiskManager):
     Evaluates the full decision chain for traceability.
     """
 
-    def approve(
+    def validate_signal(
         self,
         signal: TradeSignal,
-        signal_id: Optional[int] = None,
-        model_health: Optional[dict] = None,
-    ) -> bool:
+        market_data: pd.DataFrame,
+        open_positions: List[Dict[str, Any]],
+        model_health: Optional[Dict[str, float]] = None,
+    ) -> RiskDecision:
         """
-        Run the full 8-layer risk filter cascade.
-        Returns True only if ALL layers pass.
+        Validate a trade signal against the 8-layer cascade.
         Logs the full decision chain to the audit log.
         """
+        # We manually evaluate to capture the chain for auditing
         decision_chain = {
-            "circuit_breaker": self._check_circuit_breaker(),
-            "daily_loss": self._check_daily_loss(),
-            "max_positions": self._check_max_positions(),
-            "symbol_allocation": self._check_symbol_allocation(signal.symbol),
-            "min_confidence": self._check_minimum_confidence(signal.confidence),
+            "circuit_breaker": self._check_drawdown_breaker(),
+            "daily_loss": self.get_daily_loss_level() < 4,
+            "activity_limits": (
+                self.daily.trade_count < self.cfg.max_trades_per_day
+                and self.daily.consecutive_losses < self.cfg.max_losing_streak
+            ),
+            "exposure_limits": (
+                len(open_positions) < self.cfg.max_positions
+                and self._check_directional_exposure(signal, open_positions)
+                and self._check_total_notional(signal, open_positions, market_data)
+            ),
+            "symbol_allocation": signal.symbol == self.cfg.symbol,
+            "min_confidence": signal.confidence >= self.cfg.min_confidence,
             "risk_reward": self._check_risk_reward(signal),
-            "consecutive_losses": self._check_consecutive_losses(),
             "model_health": self._check_model_health(model_health),
         }
 
@@ -95,6 +105,13 @@ class AuditedRiskManager(RiskManager):
                     event_type="SIGNAL_REJECTED",
                     description=f"Failed filters: {reason_str}",
                     symbol=signal.symbol,
-                    signal_id=signal_id,
                 )
-        return passed
+            return RiskDecision(False, f"Failed filters: {reason_str}")
+
+        # If passed, get the adjusted lot size
+        adjusted_lots = self.size_position(signal.symbol, market_data)
+
+        if adjusted_lots < self.cfg.min_lot_size:
+            return RiskDecision(False, f"Calculated lot size {adjusted_lots} below minimum")
+
+        return RiskDecision(True, "Approved", adjusted_lots)
